# backend/docs/swagger.yaml
basePath: /
definitions:
...
  handler.batchDownloadRequest:
    properties:
      ids:
        items:
          type: string
        minItems: 1
        type: array
    required:
    - ids
    type: object
...
  models.Upload:
    properties:
      access_mode:
        type: integer
      created_at:
        type: string
      extension:
        type: string
      file_name:
        type: string
      file_path:
        type: string
      file_size:
        type: integer
      hash:
        type: string
      id:
        example: "0"
        type: string
      metadata:
        $ref: '#/definitions/models.UploadMetadata'
      mime_type:
        type: string
      status:
        $ref: '#/definitions/models.UploadStatus'
      type:
        type: string
      updated_at:
        type: string
      user_id:
        example: "0"
        type: string
    type: object
  models.UploadMetadata:
    properties:
      bucket:
        type: string
      client_ip:
        type: string
      duration:
        type: number
      extra:
        additionalProperties: {}
        type: object
      height:
        type: integer
      original_mime:
        type: string
      user_agent:
        type: string
      width:
        type: integer
    type: object
  models.UploadStatus:
    enum:
    - pending
    - used
    - deleted
    type: string
    x-enum-comments:
      UploadStatusDeleted: 已删除
      UploadStatusPending: 待使用
      UploadStatusUsed: 已使用
    x-enum-descriptions:
    - 待使用
    - 已使用
    - 已删除
    x-enum-varnames:
    - UploadStatusPending
    - UploadStatusUsed
    - UploadStatusDeleted
...
  response.Any:
    properties:
      data: {}
      error_msg:
        example: ""
        type: string
    type: object
...
info:
  contact:
    name: Wavelet
    url: https://github.com/Rain-kl/Wavelet
  description: Wavelet 平台后端 API，提供用户认证、系统配置、任务调度等通用功能。
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Wavelet API
  version: 1.0.0
paths:
...
  /api/v1/admin/uploads/download/{id}:
    get:
      description: 根据文件 ID 获取文件，以附件形式 (Attachment) 强制开启客户端浏览器下载
      parameters:
      - description: 文件 ID
        in: path
        name: id
        required: true
        type: string
      - description: 图片质量 (low, medium, high, origin)，默认为 origin
        in: query
        name: quality
        type: string
      produces:
      - application/octet-stream
      responses:
        "200":
          description: 成功下载文件
          schema:
            type: file
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.Any'
        "404":
          description: 文件不存在
          schema:
            $ref: '#/definitions/response.Any'
        "500":
          description: 服务内部错误
          schema:
            $ref: '#/definitions/response.Any'
      security:
      - SessionCookie: []
      summary: 下载单文件
      tags:
      - admin
  /api/v1/admin/uploads/download/batch:
    post:
      consumes:
      - application/json
      description: 传入多个文件 ID，后台实时将其打包压缩为 ZIP 流并输出，自动处理文件名重复冲突
      parameters:
      - description: 包含文件 ID 数组 of string 的请求体
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.batchDownloadRequest'
      produces:
      - application/octet-stream
      responses:
        "200":
          description: 成功下载打包后的 ZIP
          schema:
            type: file
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.Any'
        "500":
          description: 打包失败
          schema:
            $ref: '#/definitions/response.Any'
      security:
      - SessionCookie: []
      summary: 批量打包下载
      tags:
      - admin
...
  /api/v1/upload:
    post:
      consumes:
      - multipart/form-data
      description: 支持各种类型的通用文件上传，支持自动文件类型检测、哈希计算与“秒传”去重
      parameters:
      - description: 要上传的文件
        in: formData
        name: file
        required: true
        type: file
      - description: '业务分类 (例如: avatar, attachment, doc，默认为 generic)'
        in: formData
        name: type
        type: string
      - description: 额外的 JSON 格式元数据
        in: formData
        name: metadata
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 上传成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Any'
            - properties:
                data:
                  $ref: '#/definitions/models.Upload'
              type: object
        "400":
          description: 请求参数错误或文件受限
          schema:
            $ref: '#/definitions/response.Any'
        "401":
          description: 未登录
          schema:
            $ref: '#/definitions/response.Any'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/response.Any'
      security:
      - SessionCookie: []
      summary: 上传文件
      tags:
      - upload
...
securityDefinitions:
  SessionCookie:
    in: cookie
    name: session
    type: apiKey
swagger: "2.0"